+++ src/Column/headers/private/IteratorPrivate.h
@@ -22,25 +22,25 @@ Column myColumn;
 char *myData;
 
 // the number of bytes we get in each batch
-/*const*/ int bytesToRequest;
+/*const*/ uint64_t bytesToRequest;
 
 // the current byte offset in the column
-int curPosInColumn;
+uint64_t curPosInColumn;
 
 // the first byte we have not downloaded
-int firstInvalidByte;
+uint64_t firstInvalidByte;
 
 // tells us that the length in bytes of the current object... zero if bad
-int objLen;
+uint64_t objLen;
 
 // length in bytes of the column
-int colLength;
+uint64_t colLength;
 
 // Read or write mode
 bool isWriteOnly;
 
 // minimum bytes to get header length to read next object correctly
-int myMinByteToGetLength;
+uint64_t myMinByteToGetLength;
 
 // Do we have valid column OR column at all.
 bool isInValid;
@@ -48,11 +48,11 @@ bool isInValid;
 ///////////////// Below are variables for checkpointing ////////////////
 // Column is not needed to be checkpointed because it will be same anyway
 char *c_myData;
-//int c_bytesToRequest;
-int c_curPosInColumn;
-//int c_firstInvalidByte;
-int c_objLen;
-//int c_colLength;
+//uint64_t c_bytesToRequest;
+uint64_t c_curPosInColumn;
+//uint64_t c_firstInvalidByte;
+uint64_t c_objLen;
+//uint64_t c_colLength;
 //bool c_isWriteOnly;
-//int c_myMinByteToGetLength;
+//uint64_t c_myMinByteToGetLength;
 //bool c_isInValid;
